requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# RAG logic and Machine Learning
langchain-core
langchain-community
langchain
chromadb
sentence-transformers
pypdf
torch
transformers

# Utils
python-dotenv